requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
bitsandbytes>=0.43.0
datasets>=2.18.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
gradio>=4.20.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
